+++ src/Products/KShop.Products.Persistence/Entities/Product.cs
@@ -6,10 +6,10 @@ namespace KShop.Products.Persistence.Entities
 {
     public class Product
     {
-        public int ID { get; set; }
+        public ulong ID { get; set; }
         public string Title { get; set; }
         public decimal Price { get; set; }
-
+        
         public ICollection<ProductPosition> Positions { get; set; }
         public ICollection<ProductReserve> Reserves { get; set; }
     }
