+++ src/Products/KShop.Products.Persistence/Entities/ProductPosition.cs
@@ -2,9 +2,9 @@
 {
     public class ProductPosition
     {
-        public int ID { get; set; }
-        public int ProductID { get; set; }
-        public int Quantity { get; set; }
+        public ulong ID { get; set; }
+        public ulong ProductID { get; set; }
+        public uint Quantity { get; set; }
 
 
         public Product Product { get; set; }
